InLine_Implementation/Code/requirements.txt
cached-property==1.5.2
cycler==0.10.0
dataclasses==0.8
h5py==3.1.0
ismrmrd==1.7.2
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
Pillow==8.2.0
protobuf==3.15.8
pycparser==2.20
pydensecrf==1.0rc3
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.4.1
scipy==1.5.4
six==1.15.0
tensorboardX==2.2
torch==1.4.0
torchvision==0.5.0
xsdata==21.4
